pyproject.toml
[project]
name = "final-project-m2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arch>=7.2.0",
    "flake8>=7.2.0",
    "joblib>=1.5.1",
    "matplotlib>=3.10.3",
    "tensorflow>=2.18.0",
    "pandas>=2.2.3",
    "polars>=1.30.0",
    "pyrefly>=0.18.0",
    "pywavelets>=1.9.0",
    "scikit-learn>=1.7.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
    "tensorflow-metal>=1.2.0",
    "tensorflow-probability>=0.25.0",
    "tf-keras>=2.18.0",
    "numpy>=2.1.3",
    "esig>=1.0.0",
]
[tool.pyrefly]
project_includes = ["**/*"]
project_excludes = [
    "**/.[!/.]*",
    "**/*venv/**/*",
]
